+++ src/oracle_implementations/discount_rate/Yield/YieldValueProvider.t.sol
@@ -4,26 +4,60 @@ pragma solidity ^0.8.0;
 import "ds-test/test.sol";
 import "../../../test/utils/Caller.sol";
 import {Hevm} from "../../../test/utils/Hevm.sol";
+import {CheatCodes} from "../../../test/utils/CheatCodes.sol";
 import {Convert} from "../utils/Convert.sol";
 import {MockProvider} from "@cleanunicorn/mockprovider/src/MockProvider.sol";
 import {YieldValueProvider} from "./YieldValueProvider.sol";
 import {IYieldPool} from "./IYieldPool.sol";
 
+contract YieldValueProviderHelper is YieldValueProvider {
+    constructor(
+        // Oracle parameters
+        uint256 timeUpdateWindow_,
+        // Yield specific parameters
+        address poolAddress_,
+        uint256 maturity_,
+        int256 timeScale_,
+        uint256 timestamp_,
+        uint256 balancesRatio_
+    )
+        YieldValueProvider(
+            timeUpdateWindow_,
+            poolAddress_,
+            maturity_,
+            timeScale_,
+            timestamp_,
+            balancesRatio_
+        )
+    {
+        // Suppress warning
+        this;
+    }
+
+    function testGetValue() public returns (int256) {
+        return this.getValue();
+    }
+}
+
 contract YieldValueProviderTest is DSTest, Convert {
-    Hevm internal hevm = Hevm(DSTest.HEVM_ADDRESS);
+    CheatCodes internal cheatCodes = CheatCodes(HEVM_ADDRESS);
 
     MockProvider internal mockValueProvider;
 
     YieldValueProvider internal yieldVP;
 
     // Values taken from contract
-    // https://etherscan.io/token/0x3771c99c087a81df4633b50d8b149afaa83e3c9e
-    // at block 13911954
+    // https://etherscan.io/address/0xEf82611C6120185D3BF6e020D1993B49471E7da0
+    // at block 14834025 and block 14841073
     uint256 private _maturity = 1648177200;
     int256 private _timeScale = 3168808781; // computed from 58454204609 which is in 64.64 format
-    uint112 private _cumulativeBalancesRatio =
-        5141501570599198210548627855691773;
-    uint32 private _blockTime = 1639432842;
+    uint256 private _cumulativeBalancesRatio =
+        17436777487921011541386713127552896;
+    uint32 private _blockTime = 1653460663;
+
+    uint256 private _startBlockTimestamp = 1653370980;
+    uint256 private _startCumulativeBalanceRatio =
+        17330732019965844784580556706346119;
 
     // Default oracle parameters
     uint256 private _timeUpdateWindow = 100; // seconds
@@ -40,7 +74,9 @@ contract YieldValueProviderTest is DSTest, Convert {
             // Yield arguments
             address(mockValueProvider),
             uint256(_maturity),
-            int256(_timeScale)
+            int256(_timeScale),
+            _startBlockTimestamp,
+            _startCumulativeBalanceRatio
         );
     }
 
@@ -95,35 +131,57 @@ contract YieldValueProviderTest is DSTest, Convert {
     function test_check_cumulativeBalanceRatioLast() public {
         assertEq(
             yieldVP.cumulativeBalanceRatioLast(),
-            uconvert(_cumulativeBalancesRatio, 27, 18),
+            uconvert(_startCumulativeBalanceRatio, 27, 18),
             "Invalid cumulativeBalanceRatioLast"
         );
     }
 
-    function test_check_blockTimestampLast() public {
+    function test_check_balanceTimestampLast() public {
         assertEq(
-            yieldVP.blockTimestampLast(),
-            _blockTime,
-            "Invalid blockTimestampLast"
+            yieldVP.balanceTimestampLast(),
+            _startBlockTimestamp,
+            "Invalid balanceTimestampLast"
         );
+    }
+
+    function test_getValue_failsWhenCalledByOthers() public {
+        cheatCodes.expectRevert(
+            abi.encodeWithSelector(
+                YieldValueProvider
+                    .YieldProtocolValueProvider__getValue_onlyOracleCanCall
+                    .selector
+            )
+        );
+        yieldVP.getValue();
     }
 
     function test_getValue() public {
-        // Values take from contract
-        // https://etherscan.io/token/0x3771c99c087a81df4633b50d8b149afaa83e3c9e
-        // at block 13800244
-        int256 expectedValue = 1204540138;
+        // In order to test the getValue method we will wrap the yield value provider contract
+        // inside another helper contract witch exposes the getValue method
+        YieldValueProviderHelper yieldHelper = new YieldValueProviderHelper(
+            // Oracle arguments
+            // Time update window
+            _timeUpdateWindow,
+            // Yield arguments
+            address(mockValueProvider),
+            uint256(_maturity),
+            int256(_timeScale),
+            _startBlockTimestamp,
+            _startCumulativeBalanceRatio
+        );
+
+        int256 expectedValue = 531050252;
         setMockValues(
             mockValueProvider,
-            7342183639948751441026554744281105,
-            1640937617
+            17436777487921011541386713127552896,
+            1653460663
         );
-        int256 value = yieldVP.getValue();
-        assertTrue(value == expectedValue);
+
+        assertTrue(yieldHelper.testGetValue() == expectedValue);
     }
 
     function testFail_getValue_revertsOnOrAfterMaturityDate() public {
-        hevm.warp(yieldVP.maturity());
+        cheatCodes.warp(yieldVP.maturity());
         yieldVP.getValue();
     }
 }